refactor, rename *-int.ts files to *.int.ts
