+++ packages/ng/src/app/select/searcher/index.ts
@@ -1,3 +1,3 @@
 export {LuSelectSearcherComponent} from './select-searcher.component';
-export {LU_SELECT_SEARCH_INTL_PROVIDER, LuSelectSearchIntl} from './select-searcher-intl';
+export {LU_SELECT_SEARCH_INTL_PROVIDER, LuSelectSearchIntl} from './select-searcher.intl';
 export {LuSelectSearcherModule} from './select-searcher.module';
