+++ packages/ng/src/app/select/searcher/select-searcher.module.ts
@@ -3,7 +3,7 @@ import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
 import { LuSelectSearcherComponent } from './select-searcher.component';
-import { LU_SELECT_SEARCH_INTL_PROVIDER } from './select-searcher-intl';
+import { LU_SELECT_SEARCH_INTL_PROVIDER } from './select-searcher.intl';
 
 @NgModule({
 	imports: [
